refactor: Adapter 클릭 리스너 대신 ViewModel을 넘겨주는 방식으로 변경
