+++ app/src/main/java/com/mashup/ipdam/ui/search/adapter/kakao/PlaceViewHolder.kt
@@ -2,7 +2,6 @@ package com.mashup.ipdam.ui.search.adapter.kakao
 
 import androidx.recyclerview.widget.RecyclerView
 import com.mashup.ipdam.databinding.ItemPlaceBinding
-import com.mashup.ipdam.ui.search.adapter.PlaceAdapter
 import com.mashup.ipdam.ui.search.data.entity.kakao.Places
 
 class PlaceViewHolder(
